# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RRHH API",
        "version": "0.1"
    },
    "paths": {
...
        "/logSignaturs": {
            "get": {
                "tags": [
                    "LogSignatur"
                ],
                "summary": "getLogSignaturList",
                "description": "Get all LogSignaturs",
                "operationId": "2d618198b71744b477854338228b3fe6",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "LogSignatur"
                ],
                "summary": "createLogSignatur",
                "description": "Create LogSignatur",
                "operationId": "93e38698d551386364d8b306b147792f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/logSignaturs/{id}": {
            "get": {
                "tags": [
                    "LogSignatur"
                ],
                "summary": "getLogSignaturItem",
                "description": "Get LogSignatur",
                "operationId": "1779ea297d43aef426a30c1f845a5871",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of LogSignatur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "LogSignatur"
                ],
                "summary": "updateLogSignatur",
                "description": "Update LogSignatur",
                "operationId": "0b49f39588518a36e6d9a16cfeb0bc42",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of LogSignatur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    ""
                                ],
                                "properties": {
                                    "name": {
                                        "description": "desc",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "LogSignatur"
                ],
                "summary": "deleteLogSignatur",
                "description": "Delete LogSignatur",
                "operationId": "abec30e7a75f2736003b27de2756664d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of LogSignatur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "in": "header"
            }
        }
    }
}